+++ library/src/schemas/bigint/bigint.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 import { parse } from '../../methods/index.ts';
-import { maxRange, minRange } from '../../validations/index.ts';
+import { maxValue, minValue } from '../../validations/index.ts';
 import { bigint } from './bigint.ts';
 
 describe('bigint', () => {
@@ -20,19 +20,19 @@ describe('bigint', () => {
   });
 
   test('should execute pipe', () => {
-    const rangeError = 'Invalid range';
+    const valueError = 'Invalid value';
 
-    const schema1 = bigint([minRange(1n), maxRange(3n)]);
+    const schema1 = bigint([minValue(1n), maxValue(3n)]);
     const input1 = 2n;
     const output1 = parse(schema1, input1);
     expect(output1).toEqual(input1);
-    expect(() => parse(schema1, 0n)).toThrowError(rangeError);
-    expect(() => parse(schema1, 12n)).toThrowError(rangeError);
+    expect(() => parse(schema1, 0n)).toThrowError(valueError);
+    expect(() => parse(schema1, 12n)).toThrowError(valueError);
 
-    const schema2 = bigint('Error', [maxRange(3n)]);
+    const schema2 = bigint('Error', [maxValue(3n)]);
     const input2 = 3n;
     const output2 = parse(schema2, input2);
     expect(output2).toEqual(input2);
-    expect(() => parse(schema2, 12346789n)).toThrowError(rangeError);
+    expect(() => parse(schema2, 12346789n)).toThrowError(valueError);
   });
 });