+++ library/src/schemas/bigint/bigintAsync.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 import { parseAsync } from '../../methods/index.ts';
-import { maxRange, minRange } from '../../validations/index.ts';
+import { maxValue, minValue } from '../../validations/index.ts';
 import { bigintAsync } from './bigintAsync.ts';
 
 describe('bigintAsync', () => {
@@ -22,21 +22,21 @@ describe('bigintAsync', () => {
   });
 
   test('should execute pipe', async () => {
-    const rangeError = 'Invalid range';
+    const valueError = 'Invalid value';
 
-    const schema1 = bigintAsync([minRange(1n), maxRange(3n)]);
+    const schema1 = bigintAsync([minValue(1n), maxValue(3n)]);
     const input1 = 2n;
     const output1 = await parseAsync(schema1, input1);
     expect(output1).toEqual(input1);
-    await expect(parseAsync(schema1, 0n)).rejects.toThrowError(rangeError);
-    await expect(parseAsync(schema1, 12n)).rejects.toThrowError(rangeError);
+    await expect(parseAsync(schema1, 0n)).rejects.toThrowError(valueError);
+    await expect(parseAsync(schema1, 12n)).rejects.toThrowError(valueError);
 
-    const schema2 = bigintAsync('Error', [maxRange(3n)]);
+    const schema2 = bigintAsync('Error', [maxValue(3n)]);
     const input2 = 3n;
     const output2 = await parseAsync(schema2, input2);
     expect(output2).toEqual(input2);
     await expect(parseAsync(schema2, 12346789n)).rejects.toThrowError(
-      rangeError
+      valueError
     );
   });
 });