+++ library/src/schemas/instance/instance.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 import { parse } from '../../methods/index.ts';
-import { maxRange, minRange } from '../../validations/index.ts';
+import { maxValue, minValue } from '../../validations/index.ts';
 import { instance } from './instance.ts';
 
 describe('instance', () => {
@@ -21,28 +21,28 @@ describe('instance', () => {
   });
 
   test('should execute pipe', () => {
-    const rangeError = 'Invalid range';
+    const valueError = 'Invalid value';
 
     const schema1 = instance(Date, [
-      minRange(new Date(Date.now() - 3600000)),
-      maxRange(new Date(Date.now() + 3600000)),
+      minValue(new Date(Date.now() - 3600000)),
+      maxValue(new Date(Date.now() + 3600000)),
     ]);
     const input1 = new Date();
     const output1 = parse(schema1, input1);
     expect(output1).toEqual(input1);
     expect(() => parse(schema1, new Date(Date.now() - 4000000))).toThrowError(
-      rangeError
+      valueError
     );
     expect(() => parse(schema1, new Date(Date.now() + 4000000))).toThrowError(
-      rangeError
+      valueError
     );
 
-    const schema2 = instance(Date, 'Error', [maxRange(new Date())]);
+    const schema2 = instance(Date, 'Error', [maxValue(new Date())]);
     const input2 = new Date(Date.now() - 120000);
     const output2 = parse(schema2, input2);
     expect(output2).toEqual(input2);
     expect(() => parse(schema2, new Date(Date.now() + 1))).toThrowError(
-      rangeError
+      valueError
     );
   });
 });