+++ library/src/validations/maxValue/maxValue.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, test } from 'vitest';
-import { maxRange } from './maxRange.ts';
+import { maxValue } from './maxValue.ts';
 
-describe('maxRange', () => {
+describe('maxValue', () => {
   const info = { reason: 'any' as const };
 
-  test('should pass only valid ranges', () => {
-    const validate1 = maxRange('2023');
+  test('should pass only valid values', () => {
+    const validate1 = maxValue('2023');
     expect(validate1('', info)).toBe('');
     expect(validate1('0', info)).toBe('0');
     expect(validate1('2023', info)).toBe('2023');
@@ -13,21 +13,21 @@ describe('maxRange', () => {
     expect(() => validate1('2024', info)).toThrowError();
     expect(() => validate1('2024-01-01', info)).toThrowError();
 
-    const validate2 = maxRange(3);
+    const validate2 = maxValue(3);
     expect(validate2(-3, info)).toBe(-3);
     expect(validate2(0, info)).toBe(0);
     expect(validate2(3, info)).toBe(3);
     expect(() => validate2(4, info)).toThrowError();
     expect(() => validate2(456, info)).toThrowError();
 
-    const validate3 = maxRange(3n);
+    const validate3 = maxValue(3n);
     expect(validate3(-3n, info)).toBe(-3n);
     expect(validate3(0n, info)).toBe(0n);
     expect(validate3(3n, info)).toBe(3n);
     expect(() => validate3(4n, info)).toThrowError();
     expect(() => validate3(456n, info)).toThrowError();
 
-    const validate4 = maxRange(new Date(Date.now() + 60000));
+    const validate4 = maxValue(new Date(Date.now() + 60000));
     const date1 = new Date();
     expect(validate4(date1, info)).toEqual(date1);
     const date2 = new Date(Date.now() + 10000);
@@ -41,8 +41,8 @@ describe('maxRange', () => {
   });
 
   test('should throw custom error', () => {
-    const error = 'Value range is greater than "3"!';
-    const validate = maxRange(3, error);
+    const error = 'Value value is greater than "3"!';
+    const validate = maxValue(3, error);
     expect(() => validate(4, info)).toThrowError(error);
   });
 });