+++ library/src/validations/maxValue/maxValue.ts
@@ -2,24 +2,24 @@ import { ValiError } from '../../error/index.ts';
 import type { ValidateInfo } from '../../types.ts';
 
 /**
- * Creates a validation functions that validates the range of a string, number or date.
+ * Creates a validation functions that validates the value of a string, number or date.
  *
- * @param requirement The minimum range.
+ * @param requirement The maximum value.
  * @param error The error message.
  *
  * @returns A validation function.
  */
-export function minRange<
+export function maxValue<
   TInput extends string | number | bigint | Date,
   TRequirement extends TInput
 >(requirement: TRequirement, error?: string) {
   return (input: TInput, info: ValidateInfo) => {
-    if (input < requirement) {
+    if (input > requirement) {
       throw new ValiError([
         {
-          validation: 'min_range',
+          validation: 'max_value',
           origin: 'value',
-          message: error || 'Invalid range',
+          message: error || 'Invalid value',
           input,
           ...info,
         },
@@ -28,3 +28,10 @@ export function minRange<
     return input;
   };
 }
+
+/**
+ * See {@link maxValue}
+ *
+ * @deprecated Function has been renamed to `maxValue`.
+ */
+export const maxRange = maxValue;